+++ src/main/java/ru/job4j/design/srp/Bread.java
@@ -3,7 +3,7 @@ package ru.job4j.design.srp;
 import java.util.Date;
 
 public class Bread extends Food {
-    Bread(String name, Date createDate, Date expiryDate, Double price, Integer discount) {
+    Bread(String name, Date createDate, Date expiryDate, Double price, Double discount) {
         super(name, createDate, expiryDate, price, discount);
     }
 }
